Made transaction proof generic; extract bridge interface in typescript
