+++ tbtc-ts/src/proof.ts
@@ -1,0 +1,71 @@
+import {
+  Transaction,
+  Proof,
+  TransactionMerkleBranch,
+  Client as BitcoinClient,
+} from "./bitcoin"
+
+/**
+ * Create a proof of transaction inclusion in the block by concatenating
+ * 32-byte-long hash values. The values are converted to little endian form.
+ * @param txMerkleBranch - Branch of a merkle tree leading to a transaction.
+ * @returns Transaction inclusion proof in hexadecimal form.
+ */
+function createMerkleProof(txMerkleBranch: TransactionMerkleBranch): string {
+  let proof = Buffer.from("")
+  txMerkleBranch.merkle.forEach(function (item) {
+    proof = Buffer.concat([proof, Buffer.from(item, "hex").reverse()])
+  })
+  return proof.toString("hex")
+}
+
+/**
+ * Creates a proof that a given transaction was included in the blockchain and
+ * has accumulated the required number of confirmations.
+ * @param transactionHash - Hash of the transaction being proven.
+ * @param requiredConfirmations - Required number of confirmations.
+ * @param bitcoinClient - Bitcoin client used to interact with the network.
+ * @returns Bitcoin transaction along with the inclusion proof.
+ */
+export async function createTransactionProof(
+  transactionHash: string,
+  requiredConfirmations: number,
+  bitcoinClient: BitcoinClient
+): Promise<Transaction & Proof> {
+  const transaction = await bitcoinClient.getTransaction(transactionHash)
+  const confirmations = await bitcoinClient.getTransactionConfirmations(
+    transactionHash
+  )
+  if (confirmations < requiredConfirmations) {
+    throw new Error(
+      "Transaction confirmations number[" +
+        confirmations +
+        "] is not enough, required [" +
+        requiredConfirmations +
+        "]"
+    )
+  }
+
+  const latestBlockHeight = await bitcoinClient.latestBlockHeight()
+  const txBlockHeight = latestBlockHeight - confirmations + 1
+
+  const headersChain = await bitcoinClient.getHeadersChain(
+    txBlockHeight,
+    confirmations
+  )
+
+  const merkleBranch = await bitcoinClient.getTransactionMerkle(
+    transactionHash,
+    txBlockHeight
+  )
+
+  const merkleProof = createMerkleProof(merkleBranch)
+
+  const proof = {
+    merkleProof: merkleProof,
+    txIndexInBlock: merkleBranch.position,
+    bitcoinHeaders: headersChain,
+  }
+
+  return { ...transaction, ...proof }
+}
